Fix Windows x64 build warnings
This fixes many of the build warings identified when building
64 bit build.

+++ api/oc_blockwise.c
@@ -31,7 +31,7 @@ OC_LIST(oc_blockwise_requests);
 OC_LIST(oc_blockwise_responses);
 
 static oc_blockwise_state_t *
-oc_blockwise_init_buffer(struct oc_memb *pool, const char *href, int href_len,
+oc_blockwise_init_buffer(struct oc_memb *pool, const char *href, size_t href_len,
                          oc_endpoint_t *endpoint, oc_method_t method,
                          oc_blockwise_role_t role)
 {
@@ -103,7 +103,7 @@ oc_blockwise_response_timeout(void *data)
 }
 
 oc_blockwise_state_t *
-oc_blockwise_alloc_request_buffer(const char *href, int href_len,
+oc_blockwise_alloc_request_buffer(const char *href, size_t href_len,
                                   oc_endpoint_t *endpoint, oc_method_t method,
                                   oc_blockwise_role_t role)
 {
@@ -119,7 +119,7 @@ oc_blockwise_alloc_request_buffer(const char *href, int href_len,
 }
 
 oc_blockwise_state_t *
-oc_blockwise_alloc_response_buffer(const char *href, int href_len,
+oc_blockwise_alloc_response_buffer(const char *href, size_t href_len,
                                    oc_endpoint_t *endpoint, oc_method_t method,
                                    oc_blockwise_role_t role)
 {
@@ -262,9 +262,9 @@ oc_blockwise_find_response_buffer_by_client_cb(oc_endpoint_t *endpoint,
 #endif /* OC_CLIENT */
 
 static oc_blockwise_state_t *
-oc_blockwise_find_buffer(oc_list_t list, const char *href, int href_len,
+oc_blockwise_find_buffer(oc_list_t list, const char *href, size_t href_len,
                          oc_endpoint_t *endpoint, oc_method_t method,
-                         const char *query, int query_len,
+                         const char *query, size_t query_len,
                          oc_blockwise_role_t role)
 {
   oc_blockwise_state_t *buffer = oc_list_head(list);
@@ -282,7 +282,7 @@ oc_blockwise_find_buffer(oc_list_t list, const char *href, int href_len,
 }
 
 oc_blockwise_state_t *
-oc_blockwise_find_request_buffer(const char *href, int href_len,
+oc_blockwise_find_request_buffer(const char *href, size_t href_len,
                                  oc_endpoint_t *endpoint, oc_method_t method,
                                  const char *query, int query_len,
                                  oc_blockwise_role_t role)
@@ -292,9 +292,9 @@ oc_blockwise_find_request_buffer(const char *href, int href_len,
 }
 
 oc_blockwise_state_t *
-oc_blockwise_find_response_buffer(const char *href, int href_len,
+oc_blockwise_find_response_buffer(const char *href, size_t href_len,
                                   oc_endpoint_t *endpoint, oc_method_t method,
-                                  const char *query, int query_len,
+                                  const char *query, size_t query_len,
                                   oc_blockwise_role_t role)
 {
   return oc_blockwise_find_buffer(oc_blockwise_responses, href, href_len,
